Move scripts into backfill directory
